+++ packages/vendure-plugin-utm-tracker/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.2.0 (2025-10-30)
 
 - Added UI to view UTM parameters on the order detail page.
+- Added Time Decay Attribution model.
 
 # 1.1.1 (2025-10-27)